requirements.txt
bcrypt==3.2.0
cffi==1.14.3
pycparser==2.20
pycryptodome==3.9.8
pyfiglet==0.8.post1
six==1.15.0
